chore: tighten exception chaining, strict zip, dead int() casts
Bug-prevention nits surfaced by an extended ruff sweep (B904 + B905
+ RUF046). None of these were causing visible breakage; they are
defensive-coding gaps that the default ruff config wasn't catching.

- B904: add 'raise ... from err' to three JWTError handlers in
  attend.py so original tracebacks aren't dropped on
  invalid-token / expired-token paths.
- B905: pass strict=True to zip(images, frame_detections) in the
  legacy screen-spoof crop loop, and to zip(ens._models, outputs)
  in test_anti_spoof. Length mismatches now raise instead of silently
  truncating — surfaces real bugs rather than hiding them.
- RUF046: drop redundant int() around round() in
  crop_face_for_antispoof; round() with a single arg already returns
  int.
- Hoist local 'from app.config import settings' inside get_pipeline
  and get_anti_spoof to module-level. The local imports were leftover
  defensive code; settings is already imported at module level
  elsewhere under app/ai/ (e.g. quality.py) without circular issues.
  Local onnxruntime imports inside detector/recognizer/anti_spoof
  remain intentional — they let tests run without the optional
  runtime installed.

+++ backend/app/ai/anti_spoof.py
@@ -25,6 +25,8 @@ from pathlib import Path
 import cv2
 import numpy as np
 
+from app.config import settings
+
 logger = logging.getLogger(__name__)
 
 # Model input is 80x80 RGB, normalised the same way as ImageNet pretraining
@@ -59,10 +61,10 @@ def crop_face_for_antispoof(
 
     new_w = bw * scale
     new_h = bh * scale
-    nx1 = int(round(cx - new_w / 2.0))
-    ny1 = int(round(cy - new_h / 2.0))
-    nx2 = int(round(cx + new_w / 2.0))
-    ny2 = int(round(cy + new_h / 2.0))
+    nx1 = round(cx - new_w / 2.0)
+    ny1 = round(cy - new_h / 2.0)
+    nx2 = round(cx + new_w / 2.0)
+    ny2 = round(cy + new_h / 2.0)
 
     h, w = image.shape[:2]
     # Pad with replicated border if expanded crop falls outside the image
@@ -218,8 +220,6 @@ def get_anti_spoof() -> AntiSpoofEnsemble:
     """Lazily build and return the process-wide anti-spoof ensemble."""
     global _ensemble
     if _ensemble is None:
-        from app.config import settings
-
         _ensemble = AntiSpoofEnsemble(
             v1se_model_path=f"{settings.AI_MODEL_PATH}/MiniFASNetV1SE.onnx",
             v2_model_path=f"{settings.AI_MODEL_PATH}/MiniFASNetV2.onnx",
